==> tech_interview/genesis_lab_interview/patent/night/1_preprocessing-night.ipynb <==
import pandas as pd
import numpy as np
from sklearn.preprocessing import LabelEncoder
import warnings
warnings.filterwarnings(action='ignore')


### raw data 와 valid data 개요

raw_data = pd.read_csv("../data/night/night_sight_raw.csv", encoding='cp949')
print(raw_data.shape)
raw_data.head()

df = pd.DataFrame(raw_data, columns=raw_data.columns)

#검색식을 통해 수집한 유효특허 후보 데이터
val_data = pd.read_csv("../data/night/night_sight_val.csv", encoding = "cp949")
print(val_data.shape)
val_data.head()

#미국 특허 확인
val_data.tail()

### raw data 필요없는 column 제거
* 미국 특허만 활용
* 일본 특허 정보 x  
ex) 'Original FI[JP]', 'Original F-term[JP]', 'Original Theme Code [JP]', '출원인 식별기호[JP]'

df.columns

df = raw_data.dropna(axis=1, how='all')
df.head()

df.columns

### valid data에서 US data 외 데이터 drop

val_data = val_data[val_data.국가 == "US"]
val_data = val_data.drop_duplicates(val_data.columns)
val_data.shape
# val_data.to_csv("1.csv", encoding='cp949')

### raw data와 valid data의 공개/등록 번호 form 맞춰주기

print("<valid> \n", val_data['공개(등록)번호'].head())
print("\n<raw> \n", raw_data['공개번호/공표/재공표 '].head(), "\n", raw_data['등록번호'].head())

df.loc[:, '공개번호/공표/재공표 '] = (df.loc[:, '공개번호/공표/재공표 '].str.split('-')).str.join('')
df.loc[:, '공개번호/공표/재공표 '].fillna(value = 0, inplace=True)

# 8자리 맞추기 ex)8704653 -> 08704653
# 00000000 -> 0

df.loc[:, '등록번호'] = df.loc[:, '등록번호'].apply(lambda x:'{0:0>8}'.format(x) if(np.all(pd.notnull(x))) else x).fillna(0)

df['등록번호'].head(10)

### valid patent 확인<br>
US 유효특허 : 86개

#raw_data 중에서 val_data
y_idx_list = []
for idx, raw in enumerate(df["공개번호/공표/재공표 "]):
    for val in val_data["공개(등록)번호"]:
        if raw == val and idx not in y_idx_list:
            y_idx_list.append(idx)

for idx, raw in enumerate(df["등록번호"]):
    for val in val_data["공개(등록)번호"]:
        if raw == val and idx not in y_idx_list:
            y_idx_list.append(idx)

for idx, raw in enumerate(df["발명의 명칭"]):
    for val in val_data["발명의 명칭"]:
        if raw.lower() == val.lower() and idx not in y_idx_list:
            y_idx_list.append(idx)
y_data_list = [int(1) for i in range(len(y_idx_list))]
print(len(y_idx_list), len(y_data_list))

y_values = pd.Series(data=y_data_list, index=y_idx_list)
df.loc[:, "유효특허 여부"] = y_values
df["유효특허 여부"].fillna(0, inplace=True)

df['유효특허 여부'].value_counts()

<hr>

### IPC Digitization

df.loc[:, 'Original IPC Main'] = df.loc[:, 'Original IPC Main'].astype('str').str.slice(0,4)
df['Original IPC Main'].head()

# |로 묶인 IPC 처리
ipc_idx_list = []
ipc_all = []
for idx, i in enumerate(df['Original IPC All']):
    i = i.split('|')
    if len(i) != 1: # 여러 개 있으면
        ipc_list = []
        for ipc in i:
            ipc_list.append((ipc.strip())[:4])
        ipc_all.append(('|').join(ipc_list))
        ipc_idx_list.append(idx)
    else:
        ipc_all.append((i[0].strip())[0:4])
        ipc_idx_list.append(idx)

ipc_values = pd.Series(data = ipc_all, index=ipc_idx_list)
df.loc[:, 'Original IPC All'] = ipc_values
df['Original IPC All'].head()

### fillna

df.loc[:, '출원인'] = df.loc[:, '출원인'].fillna(value='UKN')
df[['출원인']][60:70]

### including dash/slash data preprocessing

df.loc[:, '출원일'] = (df.loc[:, '출원일'].str.split('-')).str.join('')
df.loc[:, '출원일'] = df.loc[:, '출원일'].fillna(0)

df.loc[:, '공개일'] = (df.loc[:, '공개일'].str.split('-')).str.join('')
df.loc[:, '공개일'] = df.loc[:, '공개일'].fillna(0)

df.loc[:, '등록일'] = (df.loc[:, '등록일'].str.split('-')).str.join('')
df.loc[:, '등록일'] = df.loc[:, '등록일'].fillna(0)

df.loc[:, '국제 공개번호'] = (df.loc[:, '국제 공개번호'].str.split('-')).str.join('').str.slice(2,-1)
df.loc[:, '국제 공개번호'] = (df.loc[:, '국제 공개번호'].str.split('/')).str.join('')
df.loc[:, '국제 공개번호'] = df.loc[:, '국제 공개번호'].fillna(value=int(0))

### Label Encoding

application_id = list(df['출원번호'])
le = LabelEncoder()
le.fit(application_id)

application_id_le = le.transform(application_id)
df['출원번호'] = application_id_le

patent_kind = list(df['문헌종류 코드'])
le = LabelEncoder()
le.fit(patent_kind)

patent_kind_le = le.transform(patent_kind)
df['문헌종류 코드'] = patent_kind_le

priority_id = list(df['우선권 번호'])
le = LabelEncoder()
le.fit(priority_id)

priority_id_le = le.transform(priority_id)
df['우선권 번호'] = priority_id_le

df = df.astype('str')

df.to_csv('../data/night/data_night_1.csv', encoding='cp949')
