==> tech_interview/genesis_lab_interview/patent/night/2_create_edgelist-night.ipynb <==
import pandas as pd
import numpy as np
from sklearn.preprocessing import LabelEncoder

df = pd.read_csv('../data/data_night.csv', encoding='cp949')

df = df.rename(index=str, columns={'Unnamed: 0': 'patent index'})

### ipc main to EdgeList

# ipc 한 개
main_ipc_dict = {}
main_ipc_dict['patent'] = df['patent index']
main_ipc_dict['ipc'] = df['Original IPC Main']

main_ipc_edge = pd.DataFrame(main_ipc_dict)
main_ipc_edge.to_csv("../data/night/night_main_ipc_edge.tsv", sep='\t', index=False, header=True)

### ipc all to EdgeList

# ipc 여러 개
ipc_index = []
ipc = []
for idx, i in enumerate(df['Original IPC All']):
    i = i.split('|')
    if len(i) != 1:
        for j in range(len(i)):
            ipc_index.append(idx)
            ipc.append(i[j])
    else:
        ipc_index.append(idx)
        ipc.append(i[0])

ipc_idx_values = pd.Series(data=ipc_index)
ipc_values = pd.Series(data=ipc)

all_ipc_dict = {}
all_ipc_dict['ipc'] = ipc_values
all_ipc_dict['patent'] = ipc_idx_values

all_ipc_edge = pd.DataFrame(all_ipc_dict)
all_ipc_edge.to_csv("../data/night/night_all_ipc_edge.tsv", sep='\t', index=False, header=True)

### US Class main to EdgeList

# 한 개
main_usclass_dict = {}
main_usclass_dict['patent'] = df['patent index']
main_usclass_dict['us_class'] = df['Original US Class Main[US]']

main_usclass_edge = pd.DataFrame(main_usclass_dict)
main_usclass_edge.to_csv("../data/night/night_main_usclass_edge.tsv", sep='\t', index=False, header=True)

### US Class all to EdgeList

# 여러 개
usclass_index = []
usclass = []
for idx, i in enumerate(df['Original US Class All[US]']):
    i = i.split('|')
    if len(i) != 1:
        for j in range(len(i)):
            usclass_index.append(idx)
            usclass.append(i[j].strip())
    else:
        usclass_index.append(idx)
        usclass.append(i[0])

usclass_idx_values = pd.Series(data=usclass_index)
usclass_values = pd.Series(data=usclass)

all_usclass_dict = {}
all_usclass_dict['us_class'] = usclass_values
all_usclass_dict['patent'] = usclass_idx_values

all_usclass_edge = pd.DataFrame(all_usclass_dict)
all_usclass_edge.to_csv("../data/night/night_all_usclass_edge.tsv", sep='\t', index=False, header=True)

### assignee to EdgeList

# assignee = pd.read_csv('../data/assignee.tsv', sep='\t', encoding='utf-8')
# assignee = assignee.drop('Unnamed: 0', axis=1)

# patent_id =[]
# assignee_id_list = []
# for i in range(len(assignee)):
#     assignee_id = assignee.fillna('0').iloc[[i]].values
#     for ids in assignee_id:
#         for j in ids:
#             patent_assignee = j.replace("(","").replace(")","").replace("'","").split(",")
#             if len(patent_assignee) != 1:
#                 patent_id.append(patent_assignee[0].strip())
#                 assignee_id_list.append(patent_assignee[1].strip())

# patent_values = pd.Series(data=patent_id)
# assignee_values = pd.Series(data=assignee_id_list)

# patent_assignee_dict= {}
# patent_assignee_dict['patent_id'] = patent_id
# patent_assignee_dict['assignee_id'] = assignee_id_list

# patent_assignee_edge = pd.DataFrame(patent_assignee_dict)
# patent_assignee_edge.to_csv('../data/assignee_edge.tsv', sep='\t', header=True)

### application citation to EdgeList

# application_citation = pd.read_csv('../data/application_citation.tsv', sep='\t', encoding='utf-8')
# application_citation = application_citation.drop('Unnamed: 0', axis=1)
# application_citation = application_citation.dropna(how='all')

# patent_id =[]
# application_citation_id_list = []
# for i in range(len(application_citation)):
#     application_citation_id = application_citation.fillna('0').iloc[[i]].values
#     for ids in application_citation_id:
#         for j in ids:
#             patent_application_citation = j.replace("(","").replace(")","").replace("'","").split(",")
#             if len(patent_application_citation) != 1:
#                 patent_id.append(patent_application_citation[0].strip())
#                 application_citation_id_list.append(patent_application_citation[1].strip())

# patent_values = pd.Series(data=patent_id)
# application_citation_values = pd.Series(data=application_citation_id_list)

# patent_application_citation_dict= {}
# patent_application_citation_dict['patent_id'] = patent_id
# patent_application_citation_dict['application_citation_id'] = application_citation_id_list

# patent_application_citation_edge = pd.DataFrame(patent_application_citation_dict)
# patent_application_citation_edge.to_csv('../data/application_citation_edge.tsv', sep='\t', header=True)

### grant citation to EdgeList

# grant_citation = pd.read_csv('../data/grant_citation.tsv', sep='\t', encoding='utf-8')
# grant_citation = grant_citation.drop('Unnamed: 0', axis=1)
# grant_citation = grant_citation.dropna(how='all')

# patent_id =[]
# grant_citation_id_list = []
# for i in range(len(grant_citation)):
#     grant_citation_id = grant_citation.fillna('0').iloc[[i]].values
#     for ids in grant_citation_id:
#         for j in ids:
#             patent_grant_citation = j.replace("(","").replace(")","").replace("'","").split(",")
#             if len(patent_grant_citation) != 1:
#                 patent_id.append(patent_grant_citation[0].strip())
#                 grant_citation_id_list.append(patent_grant_citation[1].strip())

# patent_values = pd.Series(data=patent_id)
# grant_citation_values = pd.Series(data=grant_citation_id_list)

# patent_grant_citation_dict= {}
# patent_grant_citation_dict['patent_id'] = patent_id
# patent_grant_citation_dict['grant_citation_id'] = grant_citation_id_list

# patent_grant_citation_edge = pd.DataFrame(patent_grant_citation_dict)
# patent_grant_citation_edge.to_csv('../data/grant_citation_edge.tsv', sep='\t', header=True)

### inventor to EdgeList

df.columns

inventor_index = []
inventor = []
for idx, i in enumerate(df['발명자/고안자']):
    i = i.split('|')
    if len(i) != 1:
        for j in range(len(i)):
            inventor_index.append(idx)
            inventor.append(i[j].strip())
    else:
        inventor_index.append(idx)
        inventor.append(i[0])

inventor_idx_values = pd.Series(data=inventor_index)
inventor_values = pd.Series(data=inventor)

inventor_dict = {}
inventor_dict['patent_id'] = inventor_idx_values
inventor_dict['inventor'] = inventor_values

inventor_edge = pd.DataFrame(inventor_dict)
inventor_edge.to_csv('../data/night/night_inventor_edge.tsv', sep='\t', encoding='utf-8', header=True)

### applicant

applicant_index = []
applicant = []
for idx, i in enumerate(df['출원인']):
    i = i.split('|')
    if len(i) != 1:
        for j in range(len(i)):
            applicant_index.append(idx)
            applicant.append(i[j].strip().lower())
    else:
        applicant_index.append(idx)
        applicant.append(i[0].lower())

applicant_idx_values = pd.Series(data=applicant_index)
applicant_values = pd.Series(data=applicant)

applicant_dict = {}
applicant_dict['patent_id'] = applicant_idx_values
applicant_dict['applicant'] = applicant_values

applicant_edge = pd.DataFrame(applicant_dict)
applicant_edge.to_csv('../data/night/night_applicant_edge.tsv', sep='\t', encoding='utf-8', header=True)
